First tests running for new server

+++ packages/test/src/storage/__tests__/Storage.test.ts
@@ -1,5 +1,5 @@
-import { describe, it, expect, beforeEach, afterEach } from "vitest";
-import { InMemoryModel, LionWebRepositoryCommunication, ServerCommunication } from "@freon4dsl/core";
+import { describe, it, expect, beforeEach, afterEach, test } from "vitest";
+import { AST, FreLogger, InMemoryModel, LionWebRepositoryCommunication, ServerCommunication2 } from "@freon4dsl/core";
 import { RulesModelEnvironment } from "../config/gen/RulesModelEnvironment.js";
 import { Data, Rules, RulesModel } from "../language/gen/index.js";
 import { fillDataUnit, fillRulesUnit, modelToString } from "./StoreModelCreator.js";
@@ -7,35 +7,46 @@ import { fillDataUnit, fillRulesUnit, modelToString } from "./StoreModelCreator.
 /**
  * For these test to run, a server should be up and running
  */
-describe.skip("Store test", () => {
+describe("Store test", () => {
     let inMemoryModel: InMemoryModel;
     let env = RulesModelEnvironment.getInstance();
-    let freonServer = new ServerCommunication();
+    let freonServer = new ServerCommunication2();
     let lionWebServer = new LionWebRepositoryCommunication();
     const communication = freonServer;
     let originalModel: RulesModel;
+    let keep = false
     // FreLogger.unmute("LionWebRepositoryCommunication")
 
     beforeEach(async () => {
+        // FreLogger.unmute("ServerCommunication2")
+        // FreLogger.unmute("InMemoryModel")
         inMemoryModel = new InMemoryModel(env, communication);
         // Create a model in the server
+
+        let unit1
+        let unit2
         originalModel = (await inMemoryModel.createModel("serverModel")) as RulesModel;
-        const unit1 = (await inMemoryModel.createUnit("dataUnit1", "Data")) as Data;
-        const unit2 = (await inMemoryModel.createUnit("rulesUnit1", "Rules")) as Rules;
-        fillDataUnit(unit1);
-        fillRulesUnit(unit2);
+        unit1 = (await inMemoryModel.createUnit("dataUnit1", "Data")) as Data;
+        unit2 = (await inMemoryModel.createUnit("rulesUnit1", "Rules")) as Rules;
+        AST.change(() => {
+            fillDataUnit(unit1);
+            fillRulesUnit(unit2);
+        })
         await inMemoryModel.saveUnit(unit1);
         await inMemoryModel.saveUnit(unit2);
     });
 
     afterEach(async () => {
-        await inMemoryModel.deleteModel();
+        if (!keep) {
+            await inMemoryModel.deleteModel();
+        }
     });
 
     it("open existing model", async () => {
         const retrievedModel = (await inMemoryModel.openModel("serverModel")) as RulesModel;
         expect(retrievedModel.getUnits().length === 2);
         expect(retrievedModel.freLanguageConcept()).toBe("RulesModel");
+        expect(retrievedModel.name).toBe("serverModel");
         expect(retrievedModel.getUnits().some((unit) => unit.name === "dataUnit1")).toBeTruthy();
         expect(retrievedModel.getUnits().some((unit) => unit.name === "rulesUnit1")).toBeTruthy();
 
@@ -57,9 +68,11 @@ describe.skip("Store test", () => {
         );
     });
 
-    it("delete unit model", async () => {
-        const unit1 = inMemoryModel.getUnitByName("dataUnit1");
-        expect(unit1).toBeDefined;
+    test("delete unit model", async () => {
+        const unit1 = await inMemoryModel.getUnitByName("dataUnit1");
+        expect(unit1).toBeDefined();
+        const unit2 = await inMemoryModel.getUnitByName("rulesUnit1");
+        expect(unit2).toBeDefined();
         expect(inMemoryModel.model.getUnits().length).toBe(2);
         await inMemoryModel.deleteUnit(unit1);
         expect(inMemoryModel.model.getUnits().length).toBe(1);
@@ -67,5 +80,20 @@ describe.skip("Store test", () => {
         const newInMemoryModel = new InMemoryModel(env, communication);
         const retrievedModel = (await newInMemoryModel.openModel("serverModel")) as RulesModel;
         expect(retrievedModel.getUnits().length === 1);
-    });
+    })
+    
+    test("rename  unit", async () => {
+        const unit1 = await inMemoryModel.getUnitByName("dataUnit1");
+        expect(unit1).toBeDefined();
+        unit1.name = "dataUnit1-changed"
+        await freonServer.renameModelUnit(inMemoryModel.model.name, "dataUnit1", "dataUnit1-changed", unit1)
+
+        const newInMemoryModel = new InMemoryModel(env, communication);
+        const retrievedModel = (await newInMemoryModel.openModel("serverModel")) as RulesModel;
+        expect(retrievedModel.getUnits().length === 2);
+        expect(retrievedModel.getUnits().map(u => u.name).includes("dataUnit1-changed"));
+        expect(retrievedModel.getUnits().map(u => u.name).includes("rulesUnit1"));
+        keep = true
+
+    })
 });
